feat: use resize observer to make flamegraph responsive with respect to the width of its parent element
