+++ projects/ngx-flamegraph/src/lib/flamegraph-node/flamegraph-node.component.html
@@ -1,5 +1,5 @@
 <svg:rect
-  (dblclick)="select()"
+  (dblclick)="zoom.emit(entry)"
   stroke="white"
   stroke-width="1px"
   pointer-events="all"
