+++ projects/ngx-flamegraph/src/lib/flamegraph-node/flamegraph-node.component.ts
@@ -13,8 +13,4 @@ export class FlamegraphNodeComponent {
   @Input() height: number;
   @Input() width: number;
   @Output() zoom = new EventEmitter();
-
-  select() {
-    this.zoom.emit(this.entry);
-  }
 }
