+++ src/app/app.component.html
@@ -1,4 +1,11 @@
 <h2>Angular Flame Graph</h2>
-<div class="wrapper">
-  <ngx-flamegraph  [config]="config" [width]="900"></ngx-flamegraph>
+
+<h3> Fixed Width</h3>
+<div class="wrapper-fixed-width">
+  <ngx-flamegraph [config]="config"></ngx-flamegraph>
+</div>
+
+<h3> Responsive </h3>
+<div class="wrapper-responsive">
+  <ngx-flamegraph [config]="config"></ngx-flamegraph>
 </div>
